Previous versions of drvAsynIPPort.c (1.29 and earlier, asyn R4-5 and earlier) attempted to allow 2 things: 1) Use an EPICS timer to time-out an I/O operation, such as send(), recv() and connect(). 2) Periodically check (every 5 seconds) during a long I/O operation to see if    the operation should be cancelled.
Item 1) above was not really implemented because there is no portable robust was
to abort a pend I/O operation.  So the timer set a flag which was checked after
the poll() was complete to see if the timeout had occured.  This was not robust,
because there were competing timers (timeout timer and poll) which could fire in
the wrong order.

Item 2) was not implemented, because asyn has no mechanism to issue a cancel
request to a driver which is blocked on an I/O operation.

Since neither of these mechanisms was working as designed, the driver has been
re-written to simplify it.  If one or both of these are to be implemented in the future
the code as of version 1.29 should be used as the starting point.

If pasynUser->timeout < 0 an infinite timeout is now used.