drvAsynIPPort: configurable disconnect on read timeout
Different EPICS records may share one TCP/IP connection (like different
axes at the same motion controller unit).

When the cable is pulled, TCP will buffer data in the send buffer,
and the readIt() function will return a timeout.

The default timeout is 2.0 seconds, and after this timeout the first axis
is put into the alarm state.

The next axis will need another 2 seconds on it's own, and for an 8-axes
system all axes are in alarm state after 16 seconds.

On the other hand, all the poll() requests are still in the send buffer,
so when the cable is put back, all 8 poll() requests are send to the
controller at once.

This may be a non-ideal situation: The controller must process many
messages at once, and a bunch of responses is send to the IOC.
The IOC should discard them all in flushIt(), or they may end up
on the wrong axis, depending on the timing.

Make it possible to close() the TCP connection already at the first timeout.
This means that all axes go into the alarm state at the same time,
and all data coming back from the controller are reliable discarded in the
TCP layer.

Re-define the parameter noProcessEos in drvAsynIPPortConfigure() into
into userFlags, which is bit wise ored:
Bit 0 is USERFLAG_NO_PROCESS_EOS
Bit 1 is USERFLAG_CLOSE_ON_READ_TIMEOUT
(All other bits are reserved, and must be set to 0)

The new feature can be activated like this:
drvAsynIPPortConfigure("MC_CPU1","192.168.1.15:5024",0,0,2)