Changed the logic so it creates maxClients drvAsynIPPort drivers at initialization rather than on-demand when clients connect.  This makes it possible to use these ports before iocInit, which is much more useful. They can be used in input and output links in EPICS records, for example. This change should be backwards compatible, since it was transparent when the ports were created.
The first drvAsynIPPort created is now named PORT_NAME:0 rather than PORT_NAME:1, where PORT_NAME is the name of the
drvAsynIPServer port created by this driver.
This might break backwards compatibility, but most clients were getting the name from a callback, so probably not.

Set the socket option SO_REUSEADDR when creating the listening socket.  Previously if a client was connected
and the IOC exited it could not be restarted again immediately.  One needed to wait for the OS to time out
the connection.  This change allows the IOC to be run again with the same listening port immediately,
without waiting for the OS timeout.

Added an exit handler to close the listen socket and delete the private structure.

Improved the report() function so it shows all the drvAsynIPPort drivers that were created.