asynRecord.: Remove warning `...comparison of unsigned enum < 0...`
In asynExceptionToString() a range check is done to make sure that
we don't read outside asynExceptionStrings[].

The current code gives a warning, since modern compiler detects
that this very enum is always >= 0 and a comparison against < 0
does not make sense,

A more clean way of accessing an array is to use a variable of type size_t.
Cast the enum into size_t before doing the range check.