Remove the path when printing file/line
When the file/line printing is enabled by the bit ASYN_TRACEINFO_SOURCE,
file names are printed with help of the __FILE__ macro.
Due to our build system, the compiler is invoked with file names
like "../devMotorAsyn.c", and this is what we find in the IOC log file later.

However, the "../" is no useful information, as it is dependig on the
make system.

Skip the "path-component" from file before the printing is done.

This will shorten the IOC log a little bit, and make it easier to read.

If someone really needs the path, we can make that configurable in
another commit in the future.