Fix arch-specific CONFIG_SITE.<host>.<target> filenames
With the original names I can't cross-build for my linux-aarch64
target from a centos-8 box because that target doesn't use TI-RPC.

The first arch-component of the name is the build host architecture,
and rarely needs configuration settings. The second component is the
target architecture being built, so any target-specific setting must
be put in a .Common.<arch> file.