Output records, asyn:READBACK=1 and late connect
When an output record is setup with PINI=0 and no VAL is specified
and asyn:READBACK=1 is set, the record will read the value from
the device and put that value into the record.

This works well when the driver has been able to read the value from the
hardware before the record in initialized.

When the driver reads the value (much) later, the normal callback/process
code will put the value into the record, but the record stays in the
"INVALID DRIVER UDF" alarm state (with UDF = 0).

Another process() will trigger the alarm processing and
the record leaves the alarm state.

For the boRecord there are 2 functions that deal with the alarm handling:
checkAlarms() and monitor()

checkAlarms() can trigger an UDF_ALARM, if the UDF field is set in the record.
monitor() posts the event to the clients.

The problem here is that checkAlarms() never resets the UDF_ALARM.
Even when the UDF field is no longer TRUE.
The UDF_ALARM will be reset "some day" by monitor() calling recGblResetAlarms().
That "some day" is actually caused by the next process, see (*).

The boRecord can be changed to handle this condition with code similar to this:

static void checkAlarms(boRecord *prec)
  {
      unsigned short val = prec->val;

      /* check for udf alarm */
      if(prec->udf == TRUE ){
          recGblSetSevr(prec,UDF_ALARM,prec->udfs);
      } else (if prec->stat == UDF_ALARM && prec->sevr == prec->udfs) {
        recGblSetSevr(prec,NO_ALARM,NO_ALARM);
      }

This handling of stat/sevr is not really in line with the rest of the code, so don't
do that for the moment.
Improve the device support instead, which already can detect the UFD=1 -> UDF=0
transition in the callback.

When UDF has been active before, start with resetting  all alarms.
New alarms may be added later in the record, depending on the record type,
the new VAL and alarm condition.

Note that change only affects output records (which may get the value from the device)

(*) Note that this process() is only triggered when a new value is read from
the driver, or the VAL field of the record changes, so it may never happen.