allow to specify an interface to bind the ip server port
The 2nd parameter of "drvAsynIPServerPortConfigure" allows to specify a
host name or ip address for multi-homed machines. If you want to specify
any interface, use the example "localhost:port", an empty host ":port" or
the any address "0.0.0.0:port". The address has to be an existing interface
on the host. If the name cannot be resolved, the call will complain and
fail. If you need the loopback interface, use "127.0.0.1:port".