Fix valueChanged flag for UInt32 paramVal.
The other paramVal types set the valueChanged flag if the parameter goes
from undefined to defined, but setUInt32 doesn't. Therefore, we need to
add a conditional to do this. Since the value is starting off as
undefined, we make sure to set data.uival to 0, since it's going to be
used to generate the callback mask as well.

This bug was detected when using setUIntDigitalParam() from the
asynPortDriver class, where setting a parameter to 0 still leaves the
corresponding PV undefined (per camonitor output), because the callback
wasn't called.

Since we are here, also:

- fix indentation to fit rest of file (4 spaces instead of 2)
- simplify application of mask: simply zero the bits which the mask has
  set, then OR the result with the masked value