Add TCP port connection timeout.
This implementation is intended to allow faster forward progress in a
startup script in cases where an unreachable device would keep an asyn
port locked until the OS timed out the connection attempt (meaning the
OS didn't have any other way of diagnosing the fact that the device was
unreachable). Forward progress for the drvAsynIPPortConfigure command
was already guaranteed, because the connection attempt happens in the
port thread, and the main thread would only wait for autoConnectTimeout
to see if the connection was successful. Since the asyn port would
remain locked, subsequent commands to the port (e.g. asynSetOption,
asynOctetSetInputEos, asynOctetSetOutputEos) would block until said OS
timeout. Now, we apply the same autoConnectTimeout to the connection
attempt itself.

This was tested by comparing to the behavior from before the patch when
connecting to "2.3.4.5:4001", where commands to the port would hang
indefinitely.

Use poll as a reasonably portable option. This takes advantage of the
existing WSAPoll and FAKE_POLL fallback implementations as well.

getsockopt(SO_ERROR) on Windows uses "int", removing the requirement for
any custom code for that platform.

We don't add any handling for EINTR, since it would add considerably
more code, and the connection can simply be retried. For that reason,
poll returning -1 is always an error.